+++ native/settings.h
@@ -1,0 +1,5 @@
+#import <Cocoa/Cocoa.h>
+NSDictionary<NSString *, NSString *> *defaultSettings(void);
+NSDictionary<NSString *, NSString *> *loadSettings(NSError **error);
+int settingsCommand(int argc, const char *argv[]);
+int showSettings(void);